Add files via upload
One issue with this write-up: I forgot to consider the case where the bottom right number is equal to 0. However, you can use the fact that the row or column connected to the bottom right is divisible by 3. Because of this, the sum of the numbers in that row/column also must be divisible by 3. You can then show from this that the bottom right number cannot be 0. Whoops.